requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
dj-database-url==0.5.0
Django==3.0.8
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
django-dynamic-formsets==0.0.8
django-heroku==0.3.1
django-widget-tweaks==1.4.8
gunicorn==20.0.4
importlib-metadata==1.7.0
Pillow==7.2.0
psycopg2==2.8.5
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
whitenoise==5.2.0
zipp==3.1.0
